+++ src/lib.rs
@@ -8,4 +8,3 @@ pub mod request;
 pub mod response;
 pub mod router;
 pub mod server;
-pub mod server_hyper;
